# AlphaZero: perf improvements for deepcopy and .to()