@adamantivm 2025-07-02
Nice and simple! LGTM! Super clear @alejandromarcu , thank you